more offsets fixes :) awesome !
Just need a few more merges and stuff to really get the `Frame` base up and running. Probably going to make a `Frame` type for better indexing in certain cases. More work should be done to keep track of these offsets in the returns of the functions, as well. This is actually surprisingly far along at this point!